Update video conversion logic in localcap.py
The method convertVideosToOpencap was modified to change how videos are moved and saved. The code no longer downloads from a remote location, but instead moves local .avi files within the server. This made import and export of metadata more efficient and it also refactored the way the session path is handled.